Rename constant_time_compare.py to util.py
